requirements.txt
h5py==2.9.0
tqdm
pandas
pytorchtools
seaborn
numpy==1.17.3 # 1.17.0
tqdm==4.19.5
matplotlib #==2.2.3
seaborn==0.9.0
scipy==1.4.1
scikit-learn==0.22.1 #0.19.1
# pip3 install torch torchvision torchaudio --extra-index-url https://download.pytorch.org/whl/cu117 # for cuda 11.7